backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
jsonschema==4.25.1
uritemplate==4.2.0



# Database driver (choose one)
mysqlclient>=2.2.4   # Uncomment if using MySQL instead

# Environment variable management
python-decouple>=3.8

# Reporting and export
pandas>=2.2
openpyxl>=3.1
reportlab>=4.2

# Optional utility libraries
django-filter>=24.3
drf-yasg>=1.21  # API documentation (Swagger)

# Security & maintenance
gunicorn>=23.0
